MODKBEKBJ-349 Make refactoring
